backend/requirements.txt
# --- CORE ---
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyMuPDF==1.25.5
python-docx==1.1.2
lxml==5.3.2
ollama==0.5.1
uvicorn==0.34.2
pandas==2.3.1
openpyxl==3.1.5

# --- GOOGLE APIs ---
google-ai-generativelanguage==0.6.15
google-generativeai==0.8.5
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-api-core==2.24.2
google-cloud-storage==2.19.0
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-aiplatform==1.92.0
google-cloud-resource-manager==1.14.2
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
grpc-google-iam-v1==0.14.2

# --- UTILITY ---
requests==2.32.3
httpx==0.28.1
httpx-sse==0.4.0
sse-starlette==2.3.4
PyYAML==6.0.2
docstring-parser==0.16
typing-extensions==4.13.2
annotated-types==0.7.0
typing-inspection==0.4.0